requirements.txt
pytest
PrettyTable
pyserial